add a temporary basepath of /mock to app-root
