+++ packages/app-root/next.config.mjs
@@ -5,6 +5,7 @@ const bundleAnalyzer = withBundleAnalyzer({
 })
 
 const nextConfig = {
+  basePath: '/mock',
   experimental: {
     optimizePackageImports: [
       '@zooniverse/react-components',
